Don't pass Config object to SpecialPageFactory
That made SpecialPageFactory potentially dependent on every
configuration option, when it only uses a few.

In the convention I introduce here, the service class does not in
principle care that its options come from configuration variables. It
just declares a list of names of options it can receive, which happen to
match the names of configuration options, and ServiceWiring just copies
them over. In the event a service wants to receive other options that
are not configuration variables, ServiceWiring can set them
appropriately from other sources. In practice, most options will
probably come from configuration variables, so using the same names and
passing them directly makes the code shorter and easier to understand.

Change-Id: I3e57513826dc5130f468486470f29eeee659d697