HistoryPager: Unset 'date-range-to' parameter when paging
Follow-up to 1b27cee8979b2fefa65dafee237485787693eb98. This fixes
paging on Special:History, and matches the existing behavior for the
previously supported 'year'/'month' parameters.

It is possible that this should be unset in IndexPager. However, in a
quick test, other pagers that have filtering by date seem to function
correctly already (tested Special:Contributions and Special:Log).
I'm making this change in HistoryPager to avoid accidentally breaking
anything, but perhaps this should be reevaluated.

Maybe the real bug is that 'date-range-to' is not ignored when
'offset' is specified?

Bug: T220781
Change-Id: I01c309178e6d5d9c919f410c64e2b833a5750141