HistoryAction: Consider 'feed=' and 'feed=0' as invalid feed actions
Previously it did a PHP truthy check, which is unrelated to the
check in question and outside its responsibility to do.

The validation of the value happens inside HistoryAction::feed
and more specifically in FeedUtils::checkFeedOutput().

Also remove needless use of unicode normalization as this only
accepts exact pre-configured presets. Those presets are linked
by the sidebar and thus never considered user or otherwise
in need of any normalization.

Change-Id: Id68a753bb5a5f03f983f5b813a42e784cf91e8c9