Cache: Make APCUBagOStuff::set return a useful value
It took me lots of time to debug a CI failure in
I9b3bc36b552901bc6ca7609ee51e80be2979a9c4. I was deceived by
$cache->set returning true, which according to the docs means that the
store was successful. But instead, this function just returns true, even
in case of failure. Make it return the result of apcu_store, which is
already in the format true = success, false = failure.

Change-Id: I2619845c12460e1acb5af696d2c11a5a4dee1bd3