tidy: Remove unused var and define $parts var to avoid undefined error
Remove unused variable $parent in RemexCompatMunger::comment(). Also,
RemexMungerData::dump() could have a possibility that all checks fail
and $parts is not defined. There are two ways we can handle this, i.e.
either by doing `$parts = []`(setting $parts to an empty array) or by
safe guarding using an `isset()` check.

This patch uses the former so that $parts is defined and can be used
below in the code.

Change-Id: I4d601a6fe36a1dce0945686cb9880336d08338be