resourceloader: Remove unused logic for WikiModule::getDB() as null
This was intended at some point for GlobalCssJs but ended up
being implemented differently once we realised that not having
any information about the remote wiki was unacceptable.

The local placeholder that points to metawiki (in case of GlobalCssJs)
must still at the very least be able to query version information
so that it gets a new version hash after changes etc. to ensure
cache rolls over.

This is being removed in preparation for refactoring where
compatibility with this broken model is no longer kept.

Change-Id: I4f1a06ace881daf1a5f202bdba0075460804de57