Gallery slideshow: Fix height calculation
Remove $container check. This check was broken due to
recent changes in skin output (adding mw-parser-output),
and was not necessary, as the computation without it is
based on $imgContainer, which is also bound by a container.

This avoids the height being set to "NaNpx", or the full
height of the page.

Bug: T196723
Change-Id: I9b4fda9c71502bf749271c55a6945d9f4a4f913e