SpecialWhatLinksHere: Conditionally hide the namespace invert checkbox
This changes hides the namespace invert checkbox like on
SpecialRecentchanges and on SpecialContributions.

Also use class="mw-input-with-label" to prevent line breaks between
checkbox and label and inside of the label.

Also use normal space (U+0020) instead of no-break space (U+00A0)
because the no-break space does not prevent any line breaks here but
gets visible at the beginning of the line in case of wrapping.

Change-Id: Id8f7201087cc35c45a4be5f457aee2b37c64511e