Allow <figure-inline> attributes through Sanitizer
Parsoid uses <figure-inline> for inline figures.  The intention is to
transition core to use <figure> and <figure-inline> as well in the
future (T118517).  As a first step (and to keep Parsoid and the legacy
parser in sync) allow <figure-inline> attributes in the Sanitizer.

Note that this does not allow <figure-inline> in wikitext,
since neither <figure> nor <figure-inline> is on the
getRecognizedTagData() list.

Bug: T51097
Bug: T118517
Bug: T118520
Change-Id: I5248717739bef0f7106c2bcf0b4a15acbc3c9a68