Log breaking change to WatchedItemQueryServiceExtension
The WatchedItemQueryServiceExtension interface was changed
in Iadbf7bc31a49 to hint against UserIdentity instead of User.
This is a breaking change that requires extensions implementing this
interface to be updated.

This once again shows why interfaces make bad extension points, see
T193613 for discussion of the topic. There just is no good way to
make this kind of change backwards-compatible.

Bug: T222677
Change-Id: Id9625824f7dec21a5b33ebf88e6f29d2a9333541