Remove unused mw.title.getDotExtension() method
Our amazing
https://codesearch.wmflabs.org/search/?q=getDotExtension&i=1
could only find a single user of this method, and that was even a
mistake, see I68426bf.

What code typically needs is:
* The full title.
* A file name without the extension.
* The extension only.

If code needs the extension with a dot, it's easier if that code either
calls getMain() or getMainText(), or prepends the dot itself.

This change helps reducing the complexity of the mw.Title module, which
is basically always loaded.

Depends-On: I68426bff662ceea8a3dd6dcae116f47f81195b4c
Change-Id: I9330eff5a73e2bccad74a78fa5b6a8897b12f666