mediawiki.Title: Remove dead code and streamline newFromUserInput()
This patch also does:
* Fix the behavior for filenames that have a dot, but no file
  extension after the dot.
* Add missing test cases.

It seems like this method is barely used anywhere:
https://codesearch.wmflabs.org/search/?q=newFromUserInput&i=1

So we might as well remove it, and inline simplified versions that
only do exactly the normalizations that are needed. For example,
silently stripping leading colons is a somewhat suprising thing to
do. Shouldn't such a title be invalid, instead of silently being
converted to something else? But this is unrelated to what this
patch does.

Change-Id: I1596fb0d9d02ca230a37f3d70d65296d6ce4ddb4