title: Convert binary regexp to use Unicode code points
The hex sequences are the raw binary values for the Unicode code points.
Now that we have a more modern PHP at hand, we can use \x{FFFF} for
Unicode characters.

I believe the /S is not needed any more. It "precompiles" the regular
expression. But this is a pretty trivial regular expression.
Precompiling it is most probably even slower.

Change-Id: I49435114b3bc31dcce8aa4e48091d509844a2a07