Don't allow completing a partial stash upload
finishUploadToStash could be called on an incomplete upload.
E.g. upload 1 chunk (that gets you the filekey) and then call
finishUploadToStash's finishUpload.

This adds another check for the stash result. The result will
only be 'Success' if all chunks have been submitted (or when
the file was uploaded as a whole, not chunked)
If we're still missing some chunks, result will be 'Continue'.

Bug: T190988
Change-Id: Ia0a4df0306b3e0cc0f2453f986b73bedd88241d3