mediawiki.Title: Remove unused signature from newFromUserInput
The signature with two parameters, the second one being an options
object, is unused:
https://codesearch.wmflabs.org/search/?q=Title.newFromUserInput&i=1

Not only that. It was broken before. This patch is not changing the
current behavior of this function! An options object passed as the
second argument was ignored before.

The method exists since I260ba88 from 2014. The alternative signature
broke with I61c7e73 from December 2018.

Arguments for not announcing this as a breaking change:
* We know it's unused (at least in code we monitor).
* It's broken for half a year, and nobody noticed.
* There is only one option anyway, "forUploading". It defaults to true,
  which is the more strict mode, and what code usually wants.

Change-Id: I6cf951d1e6f015bcd3749859c9baf002f7c82cf4