Add notice that FOR UPDATE is incompatible with aggregates
On some database types, combining locking reads with aggregate functions
is not allowed, and MediaWiki detects such cases and logs a deprecation
warning (on all database types), so we don’t really want people to use
this even if the target database type happens to support it. Add a
warning to the documentation.

Change-Id: Ib6fd2148aeb0c94c0a13b33aec08e3308c740b01